pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "passwordcrack-suite"
version = "1.0.0"
description = "Educational password cracking suite for security research and training"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Educational Use Only - See LICENSE"}
authors = [
    {name = "CyberSec Sagar Security", email = "security@example.com"}
]
keywords = ["security", "education", "password", "cracking", "research"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Education",
    "Topic :: Security",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "PySimpleGUI>=4.60.0",
    "cryptography>=41.0.0",
    "bcrypt>=4.0.0",
    "argon2-cffi>=23.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "mypy>=1.5.0",
]

gpu = [
    "pyopencl>=2023.1",
]

[project.scripts]
passwordcrack = "passwordcrack.main:main"
passwordcrack-cli = "passwordcrack.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=src/passwordcrack --cov-report=html --cov-report=term"
